modifies cjext for heap-wrapper

+++ ezlopi-util/ezlopi-util-heap/ezlopi_util_heap.h
@@ -9,8 +9,8 @@ void ezlopi_util_heap_flush(void);
 void ezlopi_util_heap_trace(void);
 
 void ezlopi_util_heap_free(void *ptr, const char * __file_name, uint32_t line_number);
-void* ezlopi_util_heap_malloc(uint32_t size, const char * file_name, uint32_t line_no);
-void* ezlopi_util_heap_calloc(uint32_t count, uint32_t size, const char * file_name, uint32_t line_no);
-void* ezlopi_util_heap_realloc(void *ptr, uint32_t new_size, const char * file_name, uint32_t line_no);
+void* ezlopi_util_heap_malloc(size_t size, const char * file_name, uint32_t line_no);
+void* ezlopi_util_heap_calloc(size_t count, size_t size, const char * file_name, uint32_t line_no);
+void* ezlopi_util_heap_realloc(void *ptr, size_t new_size, const char * file_name, uint32_t line_no);
 
 #endif // __EZLOPI_UTIL_HEAP_TRACE_H__
